third_party/WebKit/LayoutTests/fast/html/imports/import-css-sheet-in-shadow.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link id="importLink" rel="import" href="data:text/html,<div id=host></div>">
<script>
  var importDoc = importLink.import;
  var root = importDoc.querySelector("#host").attachShadow({mode:"open"});
  var shadowStyle = importDoc.createElement("style");
  shadowStyle.appendChild(importDoc.createTextNode("div{color:pink}"));
  root.appendChild(shadowStyle);

  test(() => {
    assert_true(!!shadowStyle.sheet,
        "The shadow tree Element.sheet should be defined.");
    assert_equals(shadowStyle.sheet.cssRules.length, 1,
        "The number of rules in the shadow tree stylesheet should 1.");
  }, "Element.sheet on style element in shadow tree in html import.");

  test(() => {
    assert_equals(root.styleSheets.length, 1,
        "Length of the ShadowRoot.styleSheets should be 1.");
    assert_equals(root.styleSheets[0].cssRules.length, 1,
        "The number of rules in the shadow stylesheet should be 1.");
  }, "ShadowRoot.styleSheets on shadow root in html import.");
</script>
